scripts: delete stray Linux elf2macho binary and fix format warning
scripts/elf2macho was a tracked Linux aarch64 binary left over
from a previous local build; it lives next to scripts/elf2macho.c
and confuses nobody except whoever reads the tree.  Remove it.

Also swap the %zu in scripts/elf2macho.c:1489 for %llu on
uint64_t hdr_size so the macOS clang build stops emitting a
format warning on every CI run.

+++ scripts/elf2macho.c
@@ -1486,8 +1486,8 @@ main(int argc, char **argv)
 	    24 + 32 +
 	    24 + 24 + 24 + 80 + 16 + 16;
 	if (hdr_size > PAGE_SZ) {
-		fprintf(stderr, "%s: header too large (%zu > %d)\n",
-		    argv[1], hdr_size, PAGE_SZ);
+		fprintf(stderr, "%s: header too large (%llu > %d)\n",
+		    argv[1], (unsigned long long)hdr_size, PAGE_SZ);
 		free(elf); return 1;
 	}
